Reconstruct Left Pelvis curves from Right curves

Reconstruct Left Pelvis curves from Right curves

PelR2L(data, ofs, flipped = F)

Arguments

data
dataframe with columns curve_id, side, 0,1,...,100, and whatever is specified by the ofs argument.

ofs
the column name (quoted) of data which specifies the time point of opposite foot strike.

flipped
logical. If TRUE the Right curves are flipped about the 0 degree axis. This is needed when reconstructing the coronal and transverse plane pelvis curves, but not for the saggital curves.

Value

same as data but the right curves are reconstructed to look like the left curves.

Details

Since there is only 1 pelvis for each person, the information conveyed by the left and right curve should be the same. However, the left and right curves do not typically look the same. Therefore, it might be useful to homogenize them so we are only looking at 1 set of patterns. PelR2L reconstruct the left curves from the right curves. It basically cuts out the second part of the curve, from the time of opposite (left) foot strike to the end of the gait cycle (t=100), and paste it before the original beginning of the curve (t=0:t[opposite foot off]). For the coronal and transverse plane, the whole curve can be flipped vertically about the 0 axis. i.e. multiplied by -1.

Examples

## make some fake opposite foot strike time point
kinematics$ofs=sample(58:62,nrow(kinematics),replace=T)
homogenized=PelR2L(subset(kinematics,joint=="Pel"&plane=="sag"),"ofs")

#> Joining, by = "curve_id"

matplot(t(homogenized%>%select(`0`:`100`)),type="l",lty=1,col=ifelse(temp$side=="R","blue","red"))

#> Error in ifelse(temp$side == "R", "blue", "red"): object 'temp' not found

matplot(t(subset(kinematics,joint=="Pel"&plane=="sag")%>%select(`0`:`100`)),type="l",lty=1,col=ifelse(temp$side=="R","blue","red"))

#> Error in ifelse(temp$side == "R", "blue", "red"): object 'temp' not found

PelR2L(subset(kinematics,joint=="Pel"&plane=="cor"),"ofs",flipped=TRUE)

#> Joining, by = "curve_id"

#> # A tibble: 12 x 106
#>    curve_id joint  side plane   ofs     `0`     `1`     `2`     `3`     `4`
#>       <dbl> <chr> <chr> <chr> <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1       11   Pel     L   cor    62 -2.6828 -2.4044 -2.0888 -1.7429 -1.3744
#>  2       12   Pel     R   cor    58 -0.7102 -0.3129  0.0730  0.4388  0.7755
#>  3       41   Pel     L   cor    61 -3.0019 -2.6787 -2.2869 -1.8349 -1.3319
#>  4       42   Pel     R   cor    60  1.0433  1.4483  1.8817  2.3309  2.7830
#>  5       71   Pel     L   cor    60 -7.5053 -7.0780 -6.5701 -5.9943 -5.3652
#>  6       72   Pel     R   cor    59  3.7955  4.0335  4.1979  4.2765  4.2676
#>  7      101   Pel     L   cor    62 -2.2969 -2.0101 -1.6714 -1.2893 -0.8733
#>  8      102   Pel     R   cor    62  7.5071  7.8916  8.1914  8.3997  8.5024
#>  9      131   Pel     L   cor    58 -3.7230 -3.5889 -3.3953 -3.1520 -2.8704
#> 10      132   Pel     R   cor    61 -0.9900 -0.8913 -0.8630 -0.9078 -1.0260
#> 11      161   Pel     L   cor    62 -5.0819 -4.8798 -4.5511 -4.1062 -3.5602
#> 12      162   Pel     R   cor    62  1.9137  1.9072  1.7789  1.5369  1.1935
#> # ... with 96 more variables: `5` <dbl>, `6` <dbl>, `7` <dbl>, `8` <dbl>,
#> #   `9` <dbl>, `10` <dbl>, `11` <dbl>, `12` <dbl>, `13` <dbl>, `14` <dbl>,
#> #   `15` <dbl>, `16` <dbl>, `17` <dbl>, `18` <dbl>, `19` <dbl>, `20` <dbl>,
#> #   `21` <dbl>, `22` <dbl>, `23` <dbl>, `24` <dbl>, `25` <dbl>, `26` <dbl>,
#> #   `27` <dbl>, `28` <dbl>, `29` <dbl>, `30` <dbl>, `31` <dbl>, `32` <dbl>,
#> #   `33` <dbl>, `34` <dbl>, `35` <dbl>, `36` <dbl>, `37` <dbl>, `38` <dbl>,
#> #   `39` <dbl>, `40` <dbl>, `41` <dbl>, `42` <dbl>, `43` <dbl>, `44` <dbl>,
#> #   `45` <dbl>, `46` <dbl>, `47` <dbl>, `48` <dbl>, `49` <dbl>, `50` <dbl>,
#> #   `51` <dbl>, `52` <dbl>, `53` <dbl>, `54` <dbl>, `55` <dbl>, `56` <dbl>,
#> #   `57` <dbl>, `58` <dbl>, `59` <dbl>, `60` <dbl>, `61` <dbl>, `62` <dbl>,
#> #   `63` <dbl>, `64` <dbl>, `65` <dbl>, `66` <dbl>, `67` <dbl>, `68` <dbl>,
#> #   `69` <dbl>, `70` <dbl>, `71` <dbl>, `72` <dbl>, `73` <dbl>, `74` <dbl>,
#> #   `75` <dbl>, `76` <dbl>, `77` <dbl>, `78` <dbl>, `79` <dbl>, `80` <dbl>,
#> #   `81` <dbl>, `82` <dbl>, `83` <dbl>, `84` <dbl>, `85` <dbl>, `86` <dbl>,
#> #   `87` <dbl>, `88` <dbl>, `89` <dbl>, `90` <dbl>, `91` <dbl>, `92` <dbl>,
#> #   `93` <dbl>, `94` <dbl>, `95` <dbl>, `96` <dbl>, `97` <dbl>, `98` <dbl>,
#> #   `99` <dbl>, `100` <dbl>