detect gait kinematics features

detect gait kinematics features

detectAll(x, plane = NULL, joint = NULL, comp = FALSE, fDict = NULL)

Arguments

x
data.frame which consists (at least) of columns: joint, plane, curve_id, 0,1,2,...,100

plane
a character vector. Only feature matching this plane will be detected. Whatever is specified here should appear in the fDict data.frame. For the default feature definitions, the possible choices are "Pel","Hip","Knee","Ankle","FootProgress". If NULL (default) all planes are detected.

joint
a character vector. Only feature matching this joint will be detected. If NULL (default) all joints are detected. #' Whatever is specified here should appear in the fDict data.frame. For the default feature definitions, the possible choices are "sag","cor","tra". If NULL (default) all planes are detected.

comp
logical. If TRUE the return value will have attr(.,"comp") which gives the clause result of the features.

fDict
a data.frame (at least) with columns fn,joint, and plane, used to decide which features to detect. It should have a similiar format to that generated by makeFDict. If NULL (default) the default dictionary which comes with the package is used.

Value

the same as x but augmented with extra columns indicating the detected features.

Details

For individual feature definitions see definitions.

Examples

detectAll(kinematics)

#> # A tibble: 180 x 154
#>    curve_id        joint  side plane   ofs      `0`      `1`      `2`      `3`
#>       <int>        <chr> <chr> <chr> <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1        1          Pel     L   sag    49  28.1692  28.1426  28.1362  28.1502
#>  2        2          Pel     R   sag    52  27.8176  27.8680  27.9559  28.0778
#>  3        3          Hip     L   sag    49  52.3993  52.4592  52.6038  52.7713
#>  4        4          Hip     R   sag    52  54.1121  55.2759  56.5592  57.8698
#>  5        5         Knee     L   sag    49  17.1870  19.1833  21.5705  24.1406
#>  6        6         Knee     R   sag    52  10.8145  13.8076  17.4061  21.3575
#>  7        7        Ankle     L   sag    49  -1.9163  -1.0102   0.2010   1.6512
#>  8        8        Ankle     R   sag    52  -7.2675  -7.1322  -6.3854  -5.0862
#>  9        9 FootProgress     L   sag    49 -95.2231 -94.2537 -93.2627 -92.3236
#> 10       10 FootProgress     R   sag    52 -99.2538 -97.5804 -96.0387 -94.7146
#> # ... with 170 more rows, and 145 more variables: `4` <dbl>, `5` <dbl>,
#> #   `6` <dbl>, `7` <dbl>, `8` <dbl>, `9` <dbl>, `10` <dbl>, `11` <dbl>,
#> #   `12` <dbl>, `13` <dbl>, `14` <dbl>, `15` <dbl>, `16` <dbl>, `17` <dbl>,
#> #   `18` <dbl>, `19` <dbl>, `20` <dbl>, `21` <dbl>, `22` <dbl>, `23` <dbl>,
#> #   `24` <dbl>, `25` <dbl>, `26` <dbl>, `27` <dbl>, `28` <dbl>, `29` <dbl>,
#> #   `30` <dbl>, `31` <dbl>, `32` <dbl>, `33` <dbl>, `34` <dbl>, `35` <dbl>,
#> #   `36` <dbl>, `37` <dbl>, `38` <dbl>, `39` <dbl>, `40` <dbl>, `41` <dbl>,
#> #   `42` <dbl>, `43` <dbl>, `44` <dbl>, `45` <dbl>, `46` <dbl>, `47` <dbl>,
#> #   `48` <dbl>, `49` <dbl>, `50` <dbl>, `51` <dbl>, `52` <dbl>, `53` <dbl>,
#> #   `54` <dbl>, `55` <dbl>, `56` <dbl>, `57` <dbl>, `58` <dbl>, `59` <dbl>,
#> #   `60` <dbl>, `61` <dbl>, `62` <dbl>, `63` <dbl>, `64` <dbl>, `65` <dbl>,
#> #   `66` <dbl>, `67` <dbl>, `68` <dbl>, `69` <dbl>, `70` <dbl>, `71` <dbl>,
#> #   `72` <dbl>, `73` <dbl>, `74` <dbl>, `75` <dbl>, `76` <dbl>, `77` <dbl>,
#> #   `78` <dbl>, `79` <dbl>, `80` <dbl>, `81` <dbl>, `82` <dbl>, `83` <dbl>,
#> #   `84` <dbl>, `85` <dbl>, `86` <dbl>, `87` <dbl>, `88` <dbl>, `89` <dbl>,
#> #   `90` <dbl>, `91` <dbl>, `92` <dbl>, `93` <dbl>, `94` <dbl>, `95` <dbl>,
#> #   `96` <dbl>, `97` <dbl>, `98` <dbl>, `99` <dbl>, `100` <dbl>, AIR <dbl>,
#> #   DblB <dbl>, DFdec <dbl>, ...

detectAll(kinematics, plane="sag")

#> # A tibble: 180 x 138
#>    curve_id        joint  side plane   ofs      `0`      `1`      `2`      `3`
#>       <int>        <chr> <chr> <chr> <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1        1          Pel     L   sag    49  28.1692  28.1426  28.1362  28.1502
#>  2        2          Pel     R   sag    52  27.8176  27.8680  27.9559  28.0778
#>  3        3          Hip     L   sag    49  52.3993  52.4592  52.6038  52.7713
#>  4        4          Hip     R   sag    52  54.1121  55.2759  56.5592  57.8698
#>  5        5         Knee     L   sag    49  17.1870  19.1833  21.5705  24.1406
#>  6        6         Knee     R   sag    52  10.8145  13.8076  17.4061  21.3575
#>  7        7        Ankle     L   sag    49  -1.9163  -1.0102   0.2010   1.6512
#>  8        8        Ankle     R   sag    52  -7.2675  -7.1322  -6.3854  -5.0862
#>  9        9 FootProgress     L   sag    49 -95.2231 -94.2537 -93.2627 -92.3236
#> 10       10 FootProgress     R   sag    52 -99.2538 -97.5804 -96.0387 -94.7146
#> # ... with 170 more rows, and 129 more variables: `4` <dbl>, `5` <dbl>,
#> #   `6` <dbl>, `7` <dbl>, `8` <dbl>, `9` <dbl>, `10` <dbl>, `11` <dbl>,
#> #   `12` <dbl>, `13` <dbl>, `14` <dbl>, `15` <dbl>, `16` <dbl>, `17` <dbl>,
#> #   `18` <dbl>, `19` <dbl>, `20` <dbl>, `21` <dbl>, `22` <dbl>, `23` <dbl>,
#> #   `24` <dbl>, `25` <dbl>, `26` <dbl>, `27` <dbl>, `28` <dbl>, `29` <dbl>,
#> #   `30` <dbl>, `31` <dbl>, `32` <dbl>, `33` <dbl>, `34` <dbl>, `35` <dbl>,
#> #   `36` <dbl>, `37` <dbl>, `38` <dbl>, `39` <dbl>, `40` <dbl>, `41` <dbl>,
#> #   `42` <dbl>, `43` <dbl>, `44` <dbl>, `45` <dbl>, `46` <dbl>, `47` <dbl>,
#> #   `48` <dbl>, `49` <dbl>, `50` <dbl>, `51` <dbl>, `52` <dbl>, `53` <dbl>,
#> #   `54` <dbl>, `55` <dbl>, `56` <dbl>, `57` <dbl>, `58` <dbl>, `59` <dbl>,
#> #   `60` <dbl>, `61` <dbl>, `62` <dbl>, `63` <dbl>, `64` <dbl>, `65` <dbl>,
#> #   `66` <dbl>, `67` <dbl>, `68` <dbl>, `69` <dbl>, `70` <dbl>, `71` <dbl>,
#> #   `72` <dbl>, `73` <dbl>, `74` <dbl>, `75` <dbl>, `76` <dbl>, `77` <dbl>,
#> #   `78` <dbl>, `79` <dbl>, `80` <dbl>, `81` <dbl>, `82` <dbl>, `83` <dbl>,
#> #   `84` <dbl>, `85` <dbl>, `86` <dbl>, `87` <dbl>, `88` <dbl>, `89` <dbl>,
#> #   `90` <dbl>, `91` <dbl>, `92` <dbl>, `93` <dbl>, `94` <dbl>, `95` <dbl>,
#> #   `96` <dbl>, `97` <dbl>, `98` <dbl>, `99` <dbl>, `100` <dbl>, DblB <dbl>,
#> #   DFdec <dbl>, KFxatLdec <dbl>, ...

y=detectAll(kinematics, plane="sag", joint="Ankle",comp=T)
attr(y,"comp")

#> [[1]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:mean(0:50)<1sd` DFdec
#>       <int>               <lgl> <dbl>
#>  1        7               FALSE     0
#>  2        8               FALSE     0
#>  3       37                TRUE     1
#>  4       38                TRUE     1
#>  5       67                TRUE     1
#>  6       68               FALSE     0
#>  7       97                TRUE     1
#>  8       98               FALSE     0
#>  9      127               FALSE     0
#> 10      128                TRUE     1
#> 11      157                TRUE     1
#> 12      158               FALSE     0
#> 
#> [[2]]
#> # A tibble: 12 x 4
#>    curve_id `Cl:diff(20)<-5` `Cl:c(45)<0` Desc2R
#>       <int>            <lgl>        <lgl>  <dbl>
#>  1        7            FALSE        FALSE      0
#>  2        8            FALSE        FALSE      0
#>  3       37             TRUE         TRUE      1
#>  4       38             TRUE         TRUE      1
#>  5       67             TRUE         TRUE      1
#>  6       68             TRUE        FALSE      0
#>  7       97             TRUE         TRUE      1
#>  8       98             TRUE         TRUE      1
#>  9      127             TRUE         TRUE      1
#> 10      128            FALSE        FALSE      0
#> 11      157             TRUE         TRUE      1
#> 12      158             TRUE        FALSE      0
#> 
#> [[3]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:mean(60:100)>1sd` DFatSwg
#>       <int>                 <lgl>   <dbl>
#>  1        7                 FALSE       0
#>  2        8                 FALSE       0
#>  3       37                 FALSE       0
#>  4       38                 FALSE       0
#>  5       67                 FALSE       0
#>  6       68                 FALSE       0
#>  7       97                 FALSE       0
#>  8       98                 FALSE       0
#>  9      127                 FALSE       0
#> 10      128                 FALSE       0
#> 11      157                 FALSE       0
#> 12      158                 FALSE       0
#> 
#> [[4]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:mean(80:100)<1sd`   FtD
#>       <int>                 <lgl> <dbl>
#>  1        7                  TRUE     1
#>  2        8                 FALSE     0
#>  3       37                  TRUE     1
#>  4       38                  TRUE     1
#>  5       67                  TRUE     1
#>  6       68                  TRUE     1
#>  7       97                  TRUE     1
#>  8       98                  TRUE     1
#>  9      127                  TRUE     1
#> 10      128                  TRUE     1
#> 11      157                  TRUE     1
#> 12      158                  TRUE     1
#> 
#> [[5]]
#> # A tibble: 12 x 5
#>    curve_id `Cl:.ROM(20:45)<5` `Cl:.aslope(20:45)<0.1` `Cl:c(45)<0`   H2R
#>       <int>              <lgl>                   <lgl>        <lgl> <dbl>
#>  1        7              FALSE                   FALSE        FALSE     0
#>  2        8               TRUE                    TRUE        FALSE     0
#>  3       37              FALSE                   FALSE         TRUE     0
#>  4       38              FALSE                   FALSE         TRUE     0
#>  5       67              FALSE                   FALSE         TRUE     0
#>  6       68              FALSE                   FALSE        FALSE     0
#>  7       97              FALSE                   FALSE         TRUE     0
#>  8       98              FALSE                   FALSE         TRUE     0
#>  9      127              FALSE                   FALSE         TRUE     0
#> 10      128              FALSE                   FALSE        FALSE     0
#> 11      157              FALSE                   FALSE         TRUE     0
#> 12      158              FALSE                   FALSE        FALSE     0
#> 
#> [[6]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:mean(20:45)>1sd` DFinc
#>       <int>                <lgl> <dbl>
#>  1        7                FALSE     0
#>  2        8                 TRUE     1
#>  3       37                FALSE     0
#>  4       38                FALSE     0
#>  5       67                FALSE     0
#>  6       68                 TRUE     1
#>  7       97                FALSE     0
#>  8       98                FALSE     0
#>  9      127                FALSE     0
#> 10      128                FALSE     0
#> 11      157                FALSE     0
#> 12      158                FALSE     0
#> 
#> [[7]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:max(0:60)>2sd` MaxDFinc
#>       <int>              <lgl>    <dbl>
#>  1        7              FALSE        0
#>  2        8              FALSE        0
#>  3       37              FALSE        0
#>  4       38              FALSE        0
#>  5       67              FALSE        0
#>  6       68              FALSE        0
#>  7       97              FALSE        0
#>  8       98              FALSE        0
#>  9      127              FALSE        0
#> 10      128              FALSE        0
#> 11      157              FALSE        0
#> 12      158              FALSE        0
#> 
#> [[8]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:mean(20)<1sd` PFinc
#>       <int>             <lgl> <dbl>
#>  1        7             FALSE     0
#>  2        8             FALSE     0
#>  3       37              TRUE     1
#>  4       38              TRUE     1
#>  5       67              TRUE     1
#>  6       68             FALSE     0
#>  7       97              TRUE     1
#>  8       98              TRUE     1
#>  9      127              TRUE     1
#> 10      128              TRUE     1
#> 11      157              TRUE     1
#> 12      158             FALSE     0
#> 
#> [[9]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:c(100)<2sd`   iPP
#>       <int>           <lgl> <dbl>
#>  1        7            TRUE     1
#>  2        8           FALSE     0
#>  3       37           FALSE     0
#>  4       38           FALSE     0
#>  5       67           FALSE     0
#>  6       68           FALSE     0
#>  7       97           FALSE     0
#>  8       98           FALSE     0
#>  9      127           FALSE     0
#> 10      128           FALSE     0
#> 11      157           FALSE     0
#> 12      158           FALSE     0
#> 
#> [[10]]
#> # A tibble: 12 x 3
#>    curve_id `Cl:diff(0)>0`   N1R
#>       <int>          <lgl> <dbl>
#>  1        7           TRUE     1
#>  2        8           TRUE     1
#>  3       37           TRUE     1
#>  4       38           TRUE     1
#>  5       67           TRUE     1
#>  6       68           TRUE     1
#>  7       97           TRUE     1
#>  8       98           TRUE     1
#>  9      127           TRUE     1
#> 10      128           TRUE     1
#> 11      157           TRUE     1
#> 12      158           TRUE     1
#> 
#> [[11]]
#> # A tibble: 12 x 5
#>    curve_id `Cl:.npeak(5:20)>0` `Cl:.slope(20:45)<0` `Cl:corr(0:60)>=0.8`   S2R
#>       <int>               <lgl>                <lgl>                <lgl> <dbl>
#>  1        7                TRUE                 TRUE                 TRUE     1
#>  2        8               FALSE                 TRUE                FALSE     0
#>  3       37                TRUE                 TRUE                 TRUE     1
#>  4       38                TRUE                 TRUE                 TRUE     1
#>  5       67                TRUE                 TRUE                 TRUE     1
#>  6       68                TRUE                 TRUE                 TRUE     1
#>  7       97                TRUE                 TRUE                 TRUE     1
#>  8       98                TRUE                 TRUE                 TRUE     1
#>  9      127                TRUE                 TRUE                 TRUE     1
#> 10      128                TRUE                 TRUE                 TRUE     1
#> 11      157                TRUE                 TRUE                 TRUE     1
#> 12      158                TRUE                 TRUE                 TRUE     1
#>